backend/requirements.txt
openai-whisper
llama_index
llama-index-vector-stores-lancedb
llama-index-multi-modal-llms-openai
llama-index-embeddings-clip
git+https://github.com/openai/CLIP.git
llama-index-readers-file
lancedb
moviepy
pytube
pydub
SpeechRecognition
ffmpeg-python
soundfile
torch
torchvision
matplotlib 
scikit-image
ftfy 
regex 
tqdm
yt_dlp
sentence-transformers
Pillow
fastapi
uvicorn
ffmpeg
moviepy